requirements.txt
django>=5.0
python-dotenv>=1.0
djangorestframework>=3.15
psycopg[binary]>=3.1
celery>=5.4
redis>=5.0
python-docx>=1.1
docxcompose>=1.4
pypdf>=4.2
paramiko>=3.4
PyMySQL>=1.1
